+++ src/Symfony/Bridge/Doctrine/Messenger/DoctrineTransactionMiddleware.php
@@ -13,6 +13,7 @@ namespace Symfony\Bridge\Doctrine\Messenger;
 
 use Doctrine\Common\Persistence\ManagerRegistry;
 use Doctrine\ORM\EntityManagerInterface;
+use Symfony\Component\Messenger\Envelope;
 use Symfony\Component\Messenger\Middleware\MiddlewareInterface;
 
 /**
@@ -31,7 +32,10 @@ class DoctrineTransactionMiddleware implements MiddlewareInterface
         $this->entityManagerName = $entityManagerName;
     }
 
-    public function handle($message, callable $next): void
+    /**
+     * {@inheritdoc}
+     */
+    public function handle(Envelope $envelope, callable $next): void
     {
         $entityManager = $this->managerRegistry->getManager($this->entityManagerName);
 
@@ -41,7 +45,7 @@ class DoctrineTransactionMiddleware implements MiddlewareInterface
 
         $entityManager->getConnection()->beginTransaction();
         try {
-            $next($message);
+            $next($envelope);
             $entityManager->flush();
             $entityManager->getConnection()->commit();
         } catch (\Throwable $exception) {