+++ src/Symfony/Component/Messenger/Middleware/AllowNoHandlerMiddleware.php
@@ -11,6 +11,7 @@
 
 namespace Symfony\Component\Messenger\Middleware;
 
+use Symfony\Component\Messenger\Envelope;
 use Symfony\Component\Messenger\Exception\NoHandlerForMessageException;
 
 /**
@@ -18,10 +19,13 @@ use Symfony\Component\Messenger\Exception\NoHandlerForMessageException;
  */
 class AllowNoHandlerMiddleware implements MiddlewareInterface
 {
-    public function handle($message, callable $next): void
+    /**
+     * {@inheritdoc}
+     */
+    public function handle(Envelope $envelope, callable $next): void
     {
         try {
-            $next($message);
+            $next($envelope);
         } catch (NoHandlerForMessageException $e) {
             // We allow not having a handler for this message.
         }