+++ src/Symfony/Component/Messenger/Tests/Middleware/HandleMessageMiddlewareTest.php
@@ -12,6 +12,7 @@
 namespace Symfony\Component\Messenger\Tests\Middleware;
 
 use PHPUnit\Framework\TestCase;
+use Symfony\Component\Messenger\Envelope;
 use Symfony\Component\Messenger\Handler\Locator\HandlerLocator;
 use Symfony\Component\Messenger\Middleware\HandleMessageMiddleware;
 use Symfony\Component\Messenger\Tests\Fixtures\DummyMessage;
@@ -21,6 +22,7 @@ class HandleMessageMiddlewareTest extends TestCase
     public function testItCallsTheHandlerAndNextMiddleware()
     {
         $message = new DummyMessage('Hey');
+        $envelope = new Envelope($message);
 
         $handler = $this->createPartialMock(\stdClass::class, array('__invoke'));
 
@@ -31,8 +33,8 @@ class HandleMessageMiddlewareTest extends TestCase
         )));
 
         $handler->expects($this->once())->method('__invoke')->with($message);
-        $next->expects($this->once())->method('__invoke')->with($message);
+        $next->expects($this->once())->method('__invoke')->with($envelope);
 
-        $middleware->handle($message, $next);
+        $middleware->handle($envelope, $next);
     }
 }