+++ src/Symfony/Component/Messenger/Tests/Middleware/LoggingMiddlewareTest.php
@@ -13,6 +13,7 @@ namespace Symfony\Component\Messenger\Tests\Middleware;
 
 use PHPUnit\Framework\TestCase;
 use Psr\Log\LoggerInterface;
+use Symfony\Component\Messenger\Envelope;
 use Symfony\Component\Messenger\Middleware\LoggingMiddleware;
 use Symfony\Component\Messenger\Tests\Fixtures\DummyMessage;
 
@@ -21,6 +22,7 @@ class LoggingMiddlewareTest extends TestCase
     public function testDebugLogAndNextMiddleware()
     {
         $message = new DummyMessage('Hey');
+        $envelope = new Envelope($message);
 
         $logger = $this->createMock(LoggerInterface::class);
         $logger
@@ -31,10 +33,10 @@ class LoggingMiddlewareTest extends TestCase
         $next
             ->expects($this->once())
             ->method('__invoke')
-            ->with($message)
+            ->with($envelope)
         ;
 
-        (new LoggingMiddleware($logger))->handle($message, $next);
+        (new LoggingMiddleware($logger))->handle($envelope, $next);
     }
 
     /**
@@ -43,6 +45,7 @@ class LoggingMiddlewareTest extends TestCase
     public function testWarningLogOnException()
     {
         $message = new DummyMessage('Hey');
+        $envelope = new Envelope($message);
 
         $logger = $this->createMock(LoggerInterface::class);
         $logger
@@ -57,10 +60,10 @@ class LoggingMiddlewareTest extends TestCase
         $next
             ->expects($this->once())
             ->method('__invoke')
-            ->with($message)
+            ->with($envelope)
             ->willThrowException(new \Exception())
         ;
 
-        (new LoggingMiddleware($logger))->handle($message, $next);
+        (new LoggingMiddleware($logger))->handle($envelope, $next);
     }
 }