+++ src/Symfony/Component/Messenger/CHANGELOG.md
@@ -7,6 +7,8 @@ CHANGELOG
  * The component is not experimental anymore
  * All the changes below are BC BREAKS
  * `MessageBusInterface::dispatch()` and `MiddlewareInterface::handle()` now return `void`
+ * `MiddlewareInterface::handle()` now require an `Envelope` as first argument
+ * `EnvelopeAwareInterface` has been removed
  * The signature of `Amqp*` classes changed to take a `Connection` as a first argument and an optional
    `Serializer` as a second argument.
  * `SenderLocator` has been renamed to `ContainerSenderLocator`